Update README.md
In the above code, we define a Django model `MyModel` and a signal handler function `my_signal_handler` that is connected to the `post_save` signal of `MyModel`. When an instance of `MyModel` is created (`obj = MyModel.objects.create(name="Test")`), the signal is sent and the signal handler is executed synchronously. The print statement "Object created" will only execute after the object has been created and the signal handler has completed.